% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{farm.select}
\alias{farm.select}
\title{Factor-adjusted robust model selection}
\usage{
farm.select(
  X,
  Y,
  loss = c("scad", "mcp", "lasso"),
  robust = TRUE,
  cv = FALSE,
  tau = 2,
  lin.reg = TRUE,
  K.factors = NULL,
  max.iter = 10000,
  nfolds = ceiling(length(Y)/3),
  eps = 1e-04,
  verbose = TRUE
)
}
\arguments{
\item{X}{an n x p covariate matrix with each row being a sample.}

\item{Y}{a size n outcome vector.}

\item{loss}{a character string specifying the loss function to be minimized.
Must be one of "scad", "mcp", or "lasso".}

\item{robust}{logical; whether to use robust estimators.}

\item{cv}{logical; whether to use cross-validation for robust tuning.}

\item{tau}{>0, multiplier for the tuning parameter for Huber loss.}

\item{lin.reg}{logical; TRUE for linear regression, FALSE for logistic model.}

\item{K.factors}{number of factors to be estimated. Otherwise estimated internally.}

\item{max.iter}{maximum number of iterations across the regularization path.}

\item{nfolds}{the number of cross-validation folds.}

\item{eps}{convergence threshold for \code{ncvreg}.}

\item{verbose}{logical; whether to print runtime updates.}
}
\value{
A \code{farm.select} object with selection output and prediction info.
}
\description{
Given a covariate matrix and output vector, this function first adjusts the
covariates for underlying factors and then performs penalized model selection.
The returned object can also be used for out-of-sample prediction.
}
